Move helper functions to common.c
