+++ test/c/common.h
@@ -32,3 +32,6 @@ enum TestResult {
 // post-process the returned `TestResult` to product a minimal info message for
 // the developer running the test suite.
 int run(const char *name, int (*test_function)(void));
+
+bool compare_circuits(const QkCircuit *res, const QkCircuit *expected);
+void print_circuit(const QkCircuit *qc);
