+++ test/c/test_sabre_layout.c
@@ -20,121 +20,6 @@
 #include <stdio.h>
 #include <string.h>
 
-void print_circuit(const QkCircuit *qc) {
-    size_t num_instructions = qk_circuit_num_instructions(qc);
-    QkCircuitInstruction inst;
-    for (size_t i = 0; i < num_instructions; i++) {
-        qk_circuit_get_instruction(qc, i, &inst);
-        printf("%s: qubits: (", inst.name);
-        for (uint32_t j = 0; j < inst.num_qubits; j++) {
-            printf("%d,", inst.qubits[j]);
-        }
-        printf(")");
-        uint32_t num_clbits = inst.num_clbits;
-        if (num_clbits > 0) {
-            printf(", clbits: (");
-            for (uint32_t j = 0; j < num_clbits; j++) {
-                printf("%d,", inst.clbits[j]);
-            }
-            printf(")");
-        }
-        printf("\n");
-        qk_circuit_instruction_clear(&inst);
-    }
-}
-
-bool compare_circuits(const QkCircuit *res, const QkCircuit *expected) {
-    if (qk_circuit_num_instructions(res) != qk_circuit_num_instructions(expected)) {
-        printf("Number of instructions in circuit is mismatched");
-        return false;
-    }
-    QkCircuitInstruction *res_inst = malloc(sizeof(QkCircuitInstruction));
-    QkCircuitInstruction *expected_inst = malloc(sizeof(QkCircuitInstruction));
-    for (size_t i = 0; i < qk_circuit_num_instructions(res); i++) {
-        qk_circuit_get_instruction(res, i, res_inst);
-        qk_circuit_get_instruction(expected, i, expected_inst);
-        int result = strcmp(res_inst->name, expected_inst->name);
-        if (result != 0) {
-            printf("Gate %d have different gates %s was found and expected %s", i, res_inst->name,
-                   expected_inst->name);
-            qk_circuit_instruction_clear(res_inst);
-            qk_circuit_instruction_clear(expected_inst);
-            free(res_inst);
-            free(expected_inst);
-            return false;
-        }
-        if (res_inst->num_qubits != expected_inst->num_qubits) {
-            printf("Gate %d have different number of qubits %d was found and expected %d", i,
-                   res_inst->num_qubits, expected_inst->num_qubits);
-            qk_circuit_instruction_clear(res_inst);
-            qk_circuit_instruction_clear(expected_inst);
-            free(res_inst);
-            free(expected_inst);
-            return false;
-        }
-        for (uint32_t j = 0; j < res_inst->num_qubits; j++) {
-            if (res_inst->qubits[j] != expected_inst->qubits[j]) {
-                printf("Qubit %d for gate %d are different %d was found and expected %d\n", j, i,
-                       res_inst->qubits[j] != expected_inst->qubits[j]);
-                printf("Expected circuit instructions:\n");
-                print_circuit(expected);
-                printf("Result circuit:\n");
-                print_circuit(res);
-                qk_circuit_instruction_clear(res_inst);
-                qk_circuit_instruction_clear(expected_inst);
-                free(res_inst);
-                free(expected_inst);
-                return false;
-            }
-        }
-        if (res_inst->num_clbits != expected_inst->num_clbits) {
-            printf("Gate %d have different number of clbits %d was found and expected %d", i,
-                   res_inst->num_clbits, expected_inst->num_clbits);
-            qk_circuit_instruction_clear(res_inst);
-            qk_circuit_instruction_clear(expected_inst);
-            free(res_inst);
-            free(expected_inst);
-            return false;
-        }
-        for (uint32_t j = 0; j < res_inst->num_clbits; j++) {
-            if (res_inst->clbits[j] != expected_inst->clbits[j]) {
-                printf("Clbit %d for gate %d are different %d was found and expected %d", j, i,
-                       res_inst->clbits[j] != expected_inst->clbits[j]);
-                qk_circuit_instruction_clear(res_inst);
-                qk_circuit_instruction_clear(expected_inst);
-                free(res_inst);
-                free(expected_inst);
-                return false;
-            }
-        }
-        if (res_inst->num_params != expected_inst->num_params) {
-            printf("Gate %d have different number of params %d was found and expected %d", i,
-                   res_inst->num_params, expected_inst->num_params);
-            qk_circuit_instruction_clear(res_inst);
-            qk_circuit_instruction_clear(expected_inst);
-            free(res_inst);
-            free(expected_inst);
-            return false;
-        }
-        for (uint32_t j = 0; j < res_inst->num_params; j++) {
-            if (res_inst->params[j] != expected_inst->params[j]) {
-                printf("Parameter %d for gate %d are different %d was found and expected %d", j, i,
-                       res_inst->params[j] != expected_inst->params[j]);
-                qk_circuit_instruction_clear(res_inst);
-                qk_circuit_instruction_clear(expected_inst);
-                free(res_inst);
-                free(expected_inst);
-                return false;
-            }
-        }
-        qk_circuit_instruction_clear(res_inst);
-        qk_circuit_instruction_clear(expected_inst);
-    }
-    free(res_inst);
-    free(expected_inst);
-    return true;
-}
-
 /**
  * Test running sabre layout that requires layout and routing
  */
@@ -162,7 +47,8 @@ int test_sabre_layout_applies_layout(void) {
     }
     QkSabreLayoutOptions options = qk_sabre_layout_options_default();
     options.seed = 2025;
-    QkTranspileLayout *layout_result = qk_transpiler_pass_standalone_sabre_layout(qc, target, &options);
+    QkTranspileLayout *layout_result =
+        qk_transpiler_pass_standalone_sabre_layout(qc, target, &options);
 
     QkOpCounts op_counts = qk_circuit_count_ops(qc);
     if (op_counts.len != 2) {
@@ -271,7 +157,8 @@ int test_sabre_layout_no_swap(void) {
     }
     QkSabreLayoutOptions options = qk_sabre_layout_options_default();
     options.seed = 2025;
-    QkTranspileLayout *layout_result = qk_transpiler_pass_standalone_sabre_layout(qc, target, &options);
+    QkTranspileLayout *layout_result =
+        qk_transpiler_pass_standalone_sabre_layout(qc, target, &options);
     QkCircuit *expected_circuit = qk_circuit_new(5, 0);
     for (uint32_t i = 0; i < qk_circuit_num_qubits(qc) - 1; i++) {
         uint32_t qargs[2] = {i, i + 1};
